Revert "refactor(nvim): flatten plugin folders (#19)"
This reverts commit c78cd37149fe619d3f11eff3f4c41651f543ce37.
