+++ .config/nvim/lua/plugins/coding/luasnip.lua
@@ -1,0 +1,7 @@
+return {
+  "L3MON4D3/LuaSnip",
+  opts = {
+    update_events = "TextChanged,TextChangedI",
+    delete_check_events = "TextChanged,InsertLeave",
+  },
+}
