+++ .config/nvim/lua/plugins/coding/nlsp-settings.lua
@@ -1,0 +1,7 @@
+-- https://github.com/tamago324/nlsp-settings.nvim
+
+return {
+  'tamago324/nlsp-settings.nvim',
+  cmd = 'LspSettings',
+  opts = {}
+}
