+++ .config/nvim/lua/plugins/core.lua
@@ -12,4 +12,9 @@ return {
       },
     },
   },
+  { import = "plugins.coding" },
+  { import = "plugins.editor" },
+  { import = "plugins.util" },
+  { import = "plugins.syntax" },
+  pcall(require, "keymaps") and require("keymaps").plugins or nil,
 }
