+++ .config/nvim/lua/plugins/editor/dressing.lua
@@ -1,0 +1,26 @@
+-- https://github.com/stevearc/dressing.nvim
+
+local float = require("util.opts").float
+
+return {
+  "stevearc/dressing.nvim",
+  opts = {
+    input = {
+      win_options = {
+        winblend = float.winblend,
+      },
+    },
+    select = {
+      nui = {
+        win_options = {
+          winblend = float.winblend,
+        },
+      },
+    },
+    builtin = {
+      win_options = {
+        winblend = float.winblend,
+      },
+    },
+  },
+}
