+++ .config/nvim/lua/plugins/editor/indent-blankline.lua
@@ -1,0 +1,8 @@
+-- https://github.com/lukas-reineke/indent-blankline.nvim
+
+return {
+  "lukas-reineke/indent-blankline.nvim",
+  opts = {
+    filetype_exclude = { "bigfile" },
+  },
+}
