+++ .config/nvim/lua/plugins/editor/markdown-preview.lua
@@ -1,0 +1,8 @@
+return {
+  "iamcco/markdown-preview.nvim",
+  build = function()
+    vim.fn["mkdp#util#install"]()
+  end,
+  pin = true,
+  ft = { "markdown", "md" },
+}
