+++ .config/nvim/lua/plugins/editor/mini-indentscope.lua
@@ -1,0 +1,12 @@
+-- https://github.com/echasnovski/mini.indentscope
+
+return {
+  "echasnovski/mini.indentscope",
+  lazy = true,
+  opts = function(_, opts)
+    opts.draw = {
+      animation = require("mini.indentscope").gen_animation.none(),
+    }
+    return opts
+  end,
+}
