+++ .config/nvim/lua/plugins/editor/spectre.lua
@@ -1,0 +1,6 @@
+-- navigation
+-- https://github.com/nvim-pack/nvim-spectre
+
+return {
+  "nvim-pack/nvim-spectre",
+}
