+++ .config/nvim/lua/plugins/editor/todo-comments.lua
@@ -1,0 +1,34 @@
+-- https://github.com/folke/todo-comments.nvim
+
+local icons = require("config.icons")
+local icons_ui = icons.get("ui", { "space" })
+local icons_diagnostics = icons.get("diagnostics", { "space" })
+
+return {
+  "folke/todo-comments.nvim",
+  opts = {
+    keywords = {
+      FIX = {
+        icon = icons_ui.Bug,
+      },
+      TODO = {
+        icon = icons_ui.Check,
+      },
+      HACK = {
+        icon = icons_ui.Skull,
+      },
+      WARN = {
+        icon = icons_diagnostics.Warn,
+      },
+      PERF = {
+        icon = icons_ui.Dashboard,
+      },
+      NOTE = {
+        icon = icons_ui.Comment,
+      },
+      TEST = {
+        icon = icons_ui.Beaker,
+      },
+    },
+  },
+}
