+++ .config/nvim/lua/plugins/editor/trouble.lua
@@ -1,0 +1,6 @@
+-- https://github.com/folke/trouble.nvim
+
+return {
+  "folke/trouble.nvim",
+  opts = {},
+}
