+++ .config/nvim/lua/plugins/util/colorizer.lua
@@ -1,0 +1,13 @@
+-- https://github.com/norcalli/nvim-colorizer.lua
+
+return {
+  "NvChad/nvim-colorizer.lua",
+  lazy = true,
+  event = { "BufReadPost" },
+  opts = {
+    user_default_options = {
+      RRGGBBAA = true,
+      tailwind = true,
+    },
+  },
+}
