+++ .config/nvim/lua/plugins/util/project.lua
@@ -1,0 +1,3 @@
+return {
+  { import = "lazyvimp.project" },
+}
