+++ .config/nvim/lua/plugins/util/wakatime.lua
@@ -1,0 +1,4 @@
+return {
+  "wakatime/vim-wakatime",
+  event = "BufReadPost",
+}
